requirements.txt
langchain-openai>=0.0.2
python-dotenv>=1.0.0
PyGithub>=2.1.1
mcp>=0.1.0 
Flask